api/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.42.0
python-dotenv==1.2.2
asyncpg==0.31.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
argon2-cffi==25.1.0
bcrypt==5.0.0
python-multipart==0.0.22
apscheduler==3.10.4
openpyxl==3.1.5
xlrd==2.0.1
aiofiles==25.1.0
aiosmtplib==5.1.0
slowapi==0.1.9
psutil==7.2.2
fpdf2==2.8.7
httpx==0.28.1
